:recycle: Address final review comments Move types to respective file to reduce clutter Organise base and p2p code for readability
