+++ elements/lisk-p2p/src/peer_pool.ts
@@ -17,9 +17,10 @@
  * The purpose of the PeerPool is to provide a simple interface for selecting,
  * interacting with and handling aggregated events from a collection of peers.
  */
-import { codec } from '@liskhq/lisk-codec';
 import { EventEmitter } from 'events';
 import { SCServerSocket } from 'socketcluster-server';
+import { codec } from '@liskhq/lisk-codec';
+
 import {
 	ConnectionKind,
 	DEFAULT_LOCALHOST_IP,
@@ -51,7 +52,7 @@ import {
 	REMOTE_EVENT_POST_NODE_INFO,
 } from './events';
 import { P2PRequest } from './p2p_request';
-import { ConnectionState, InboundPeer, OutboundPeer, Peer, PeerConfig } from './peer';
+import { ConnectionState, InboundPeer, OutboundPeer, Peer } from './peer';
 import { PeerBook } from './peer_book/peer_book';
 import { validatePayloadSize } from './utils';
 import {
@@ -68,12 +69,12 @@ import {
 	RPCSchemas,
 	P2PRequestPacketBufferData,
 	P2PMessagePacketBufferData,
+	PeerConfig,
+	PeerPoolConfig,
 } from './types';
 import { encodeNodeInfo } from './utils/codec';
 // eslint-disable-next-line import/order
 import shuffle = require('lodash.shuffle');
-
-// eslint-disable-next-line @typescript-eslint/no-require-imports
 
 interface FilterPeersOptions {
 	readonly category: PROTECTION_CATEGORY;
@@ -114,34 +115,6 @@ export enum PROTECTION_CATEGORY {
 	LATENCY = 'latency',
 	RESPONSE_RATE = 'responseRate',
 	CONNECT_TIME = 'connectTime',
-}
-
-export interface PeerPoolConfig {
-	readonly hostPort: number;
-	readonly ackTimeout?: number;
-	readonly connectTimeout?: number;
-	readonly wsMaxPayload?: number;
-	readonly maxPeerInfoSize: number;
-	readonly peerSelectionForSend: P2PPeerSelectionForSendFunction;
-	readonly peerSelectionForRequest: P2PPeerSelectionForRequestFunction;
-	readonly peerSelectionForConnection: P2PPeerSelectionForConnectionFunction;
-	readonly sendPeerLimit: number;
-	readonly peerBanTime: number;
-	readonly maxOutboundConnections: number;
-	readonly maxInboundConnections: number;
-	readonly maxPeerDiscoveryResponseLength: number;
-	readonly outboundShuffleInterval: number;
-	readonly netgroupProtectionRatio: number;
-	readonly latencyProtectionRatio: number;
-	readonly productivityProtectionRatio: number;
-	readonly longevityProtectionRatio: number;
-	readonly wsMaxMessageRate: number;
-	readonly wsMaxMessageRatePenalty: number;
-	readonly rateCalculationInterval: number;
-	readonly peerStatusMessageRate: number;
-	readonly secret: number;
-	readonly peerBook: PeerBook;
-	readonly rpcSchemas: RPCSchemas;
 }
 
 export class PeerPool extends EventEmitter {
